requirements.txt
flask
python-telegram-bot==13.15
replicate
requests